src/main/java/ru/skypro/homework/mapper/CreateOrUpdateCommentMapper.java
package ru.skypro.homework.mapper;

import org.mapstruct.Mapper;
import org.mapstruct.Mapping;
import ru.skypro.homework.dto.CreateOrUpdateComment;
import ru.skypro.homework.model.Comment;

@Mapper(componentModel = "spring")
public abstract class CreateOrUpdateCommentMapper {

    @Mapping(target = "id", ignore = true)
    @Mapping(target = "createdAt", ignore = true)
    @Mapping(target = "user", ignore = true)
    public abstract Comment toComment(CreateOrUpdateComment createOrUpdateComment);
}
